Consolidate NativeRuntimeEventSource and PortableThreadPoolEventSource
