+++ src/libraries/System.Private.CoreLib/src/System/Threading/PortableThreadPool.WaitThread.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Diagnostics;
+using System.Diagnostics.Tracing;
+
 using Microsoft.Win32.SafeHandles;
 
 namespace System.Threading
@@ -36,9 +38,9 @@ namespace System.Threading
         /// <param name="handle">A description of the requested registration.</param>
         internal void RegisterWaitHandle(RegisteredWaitHandle handle)
         {
-            if (PortableThreadPoolEventSource.Log.IsEnabled())
+            if (NativeRuntimeEventSource.Log.IsEnabled())
             {
-                PortableThreadPoolEventSource.Log.ThreadPoolIOEnqueue(handle);
+                NativeRuntimeEventSource.Log.ThreadPoolIOEnqueue(handle);
             }
 
             _waitThreadLock.Acquire();
@@ -75,9 +77,9 @@ namespace System.Threading
 
         internal static void CompleteWait(RegisteredWaitHandle handle, bool timedOut)
         {
-            if (PortableThreadPoolEventSource.Log.IsEnabled())
+            if (NativeRuntimeEventSource.Log.IsEnabled())
             {
-                PortableThreadPoolEventSource.Log.ThreadPoolIODequeue(handle);
+                NativeRuntimeEventSource.Log.ThreadPoolIODequeue(handle);
             }
 
             handle.PerformCallback(timedOut);
