--- src/libraries/System.Private.CoreLib/src/System/Threading/PortableThreadPoolEventSource.cs
@@ -1,336 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Diagnostics.Tracing;
-using System.Runtime.CompilerServices;
-using Internal.Runtime.CompilerServices;
-
-namespace System.Threading
-{
-    // Currently with EventPipe there isn't a way to move events from the native side to the managed side and get the same
-    // experience. For now, the same provider name and guid are used as the native side and a temporary change has been made to
-    // EventPipe in CoreCLR to get thread pool events in performance profiles when the portable thread pool is enabled, as that
-    // seems to be the easiest way currently and the closest to the experience when the portable thread pool is disabled.
-    // TODO: Long-term options (also see https://github.com/dotnet/runtime/issues/38763):
-    // - Use NativeRuntimeEventSource instead, change its guid to match the provider guid from the native side, and fix the
-    //   underlying issues such that duplicate events are not sent. This should get the same experience as sending events from
-    //   the native side, and would allow easily moving other events from the native side to the managed side in the future if
-    //   necessary.
-    // - Use a different provider name and guid (maybe "System.Threading.ThreadPool"), update PerfView and dotnet-trace to
-    //   enable the provider by default when the Threading or other ThreadPool-related keywords are specified for the runtime
-    //   provider, and update PerfView with a trace event parser for the new provider so that it knows about the events and may
-    //   use them to identify thread pool threads.
-    [EventSource(Name = "Microsoft-Windows-DotNETRuntime", Guid = "e13c0d23-ccbc-4e12-931b-d9cc2eee27e4")]
-    internal sealed class PortableThreadPoolEventSource : EventSource
-    {
-        // This value does not seem to be used, leaving it as zero for now. It may be useful for a scenario that may involve
-        // multiple instances of the runtime within the same process, but then it seems unlikely that both instances' thread
-        // pools would be in moderate use.
-        private const ushort DefaultClrInstanceId = 0;
-
-        private static class Messages
-        {
-            public const string WorkerThread = "ActiveWorkerThreadCount={0};\nRetiredWorkerThreadCount={1};\nClrInstanceID={2}";
-            public const string WorkerThreadAdjustmentSample = "Throughput={0};\nClrInstanceID={1}";
-            public const string WorkerThreadAdjustmentAdjustment = "AverageThroughput={0};\nNewWorkerThreadCount={1};\nReason={2};\nClrInstanceID={3}";
-            public const string WorkerThreadAdjustmentStats = "Duration={0};\nThroughput={1};\nThreadWave={2};\nThroughputWave={3};\nThroughputErrorEstimate={4};\nAverageThroughputErrorEstimate={5};\nThroughputRatio={6};\nConfidence={7};\nNewControlSetting={8};\nNewThreadWaveMagnitude={9};\nClrInstanceID={10}";
-            public const string IOEnqueue = "NativeOverlapped={0};\nOverlapped={1};\nMultiDequeues={2};\nClrInstanceID={3}";
-            public const string IO = "NativeOverlapped={0};\nOverlapped={1};\nClrInstanceID={2}";
-            public const string WorkingThreadCount = "Count={0};\nClrInstanceID={1}";
-        }
-
-        // The task definitions for the ETW manifest
-        public static class Tasks // this name and visibility is important for EventSource
-        {
-            public const EventTask ThreadPoolWorkerThread = (EventTask)16;
-            public const EventTask ThreadPoolWorkerThreadAdjustment = (EventTask)18;
-            public const EventTask ThreadPool = (EventTask)23;
-            public const EventTask ThreadPoolWorkingThreadCount = (EventTask)22;
-        }
-
-        public static class Opcodes // this name and visibility is important for EventSource
-        {
-            public const EventOpcode IOEnqueue = (EventOpcode)13;
-            public const EventOpcode IODequeue = (EventOpcode)14;
-            public const EventOpcode Wait = (EventOpcode)90;
-            public const EventOpcode Sample = (EventOpcode)100;
-            public const EventOpcode Adjustment = (EventOpcode)101;
-            public const EventOpcode Stats = (EventOpcode)102;
-        }
-
-        public static class Keywords // this name and visibility is important for EventSource
-        {
-            public const EventKeywords ThreadingKeyword = (EventKeywords)0x10000;
-            public const EventKeywords ThreadTransferKeyword = (EventKeywords)0x80000000;
-        }
-
-        public enum ThreadAdjustmentReasonMap : uint
-        {
-            Warmup,
-            Initializing,
-            RandomMove,
-            ClimbingMove,
-            ChangePoint,
-            Stabilizing,
-            Starvation,
-            ThreadTimedOut
-        }
-
-        private PortableThreadPoolEventSource()
-            : base(
-                  new Guid(0xe13c0d23, 0xccbc, 0x4e12, 0x93, 0x1b, 0xd9, 0xcc, 0x2e, 0xee, 0x27, 0xe4),
-                  "Microsoft-Windows-DotNETRuntime")
-        {
-        }
-
-        [NonEvent]
-        private unsafe void WriteThreadEvent(int eventId, uint numExistingThreads)
-        {
-            uint retiredWorkerThreadCount = 0;
-            ushort clrInstanceId = DefaultClrInstanceId;
-
-            EventData* data = stackalloc EventData[3];
-            data[0].DataPointer = (IntPtr)(&numExistingThreads);
-            data[0].Size = sizeof(uint);
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&retiredWorkerThreadCount);
-            data[1].Size = sizeof(uint);
-            data[1].Reserved = 0;
-            data[2].DataPointer = (IntPtr)(&clrInstanceId);
-            data[2].Size = sizeof(ushort);
-            data[2].Reserved = 0;
-            WriteEventCore(eventId, 3, data);
-        }
-
-        [Event(50, Level = EventLevel.Informational, Message = Messages.WorkerThread, Task = Tasks.ThreadPoolWorkerThread, Opcode = EventOpcode.Start, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        public unsafe void ThreadPoolWorkerThreadStart(
-            uint ActiveWorkerThreadCount,
-            uint RetiredWorkerThreadCount = 0,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (IsEnabled(EventLevel.Informational, Keywords.ThreadingKeyword))
-            {
-                WriteThreadEvent(50, ActiveWorkerThreadCount);
-            }
-        }
-
-        [Event(51, Level = EventLevel.Informational, Message = Messages.WorkerThread, Task = Tasks.ThreadPoolWorkerThread, Opcode = EventOpcode.Stop, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        public void ThreadPoolWorkerThreadStop(
-            uint ActiveWorkerThreadCount,
-            uint RetiredWorkerThreadCount = 0,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (IsEnabled(EventLevel.Informational, Keywords.ThreadingKeyword))
-            {
-                WriteThreadEvent(51, ActiveWorkerThreadCount);
-            }
-        }
-
-        [Event(57, Level = EventLevel.Informational, Message = Messages.WorkerThread, Task = Tasks.ThreadPoolWorkerThread, Opcode = Opcodes.Wait, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        public void ThreadPoolWorkerThreadWait(
-            uint ActiveWorkerThreadCount,
-            uint RetiredWorkerThreadCount = 0,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (IsEnabled(EventLevel.Informational, Keywords.ThreadingKeyword))
-            {
-                WriteThreadEvent(57, ActiveWorkerThreadCount);
-            }
-        }
-
-        [Event(54, Level = EventLevel.Informational, Message = Messages.WorkerThreadAdjustmentSample, Task = Tasks.ThreadPoolWorkerThreadAdjustment, Opcode = Opcodes.Sample, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        public unsafe void ThreadPoolWorkerThreadAdjustmentSample(
-            double Throughput,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (!IsEnabled(EventLevel.Informational, Keywords.ThreadingKeyword))
-            {
-                return;
-            }
-
-            EventData* data = stackalloc EventData[2];
-            data[0].DataPointer = (IntPtr)(&Throughput);
-            data[0].Size = sizeof(double);
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&ClrInstanceID);
-            data[1].Size = sizeof(ushort);
-            data[1].Reserved = 0;
-            WriteEventCore(54, 2, data);
-        }
-
-        [Event(55, Level = EventLevel.Informational, Message = Messages.WorkerThreadAdjustmentAdjustment, Task = Tasks.ThreadPoolWorkerThreadAdjustment, Opcode = Opcodes.Adjustment, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        public unsafe void ThreadPoolWorkerThreadAdjustmentAdjustment(
-            double AverageThroughput,
-            uint NewWorkerThreadCount,
-            ThreadAdjustmentReasonMap Reason,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (!IsEnabled(EventLevel.Informational, Keywords.ThreadingKeyword))
-            {
-                return;
-            }
-
-            EventData* data = stackalloc EventData[4];
-            data[0].DataPointer = (IntPtr)(&AverageThroughput);
-            data[0].Size = sizeof(double);
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&NewWorkerThreadCount);
-            data[1].Size = sizeof(uint);
-            data[1].Reserved = 0;
-            data[2].DataPointer = (IntPtr)(&Reason);
-            data[2].Size = sizeof(ThreadAdjustmentReasonMap);
-            data[2].Reserved = 0;
-            data[3].DataPointer = (IntPtr)(&ClrInstanceID);
-            data[3].Size = sizeof(ushort);
-            data[3].Reserved = 0;
-            WriteEventCore(55, 4, data);
-        }
-
-        [Event(56, Level = EventLevel.Verbose, Message = Messages.WorkerThreadAdjustmentStats, Task = Tasks.ThreadPoolWorkerThreadAdjustment, Opcode = Opcodes.Stats, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        public unsafe void ThreadPoolWorkerThreadAdjustmentStats(
-            double Duration,
-            double Throughput,
-            double ThreadWave,
-            double ThroughputWave,
-            double ThroughputErrorEstimate,
-            double AverageThroughputErrorEstimate,
-            double ThroughputRatio,
-            double Confidence,
-            double NewControlSetting,
-            ushort NewThreadWaveMagnitude,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (!IsEnabled(EventLevel.Verbose, Keywords.ThreadingKeyword))
-            {
-                return;
-            }
-
-            EventData* data = stackalloc EventData[11];
-            data[0].DataPointer = (IntPtr)(&Duration);
-            data[0].Size = sizeof(double);
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&Throughput);
-            data[1].Size = sizeof(double);
-            data[1].Reserved = 0;
-            data[2].DataPointer = (IntPtr)(&ThreadWave);
-            data[2].Size = sizeof(double);
-            data[2].Reserved = 0;
-            data[3].DataPointer = (IntPtr)(&ThroughputWave);
-            data[3].Size = sizeof(double);
-            data[3].Reserved = 0;
-            data[4].DataPointer = (IntPtr)(&ThroughputErrorEstimate);
-            data[4].Size = sizeof(double);
-            data[4].Reserved = 0;
-            data[5].DataPointer = (IntPtr)(&AverageThroughputErrorEstimate);
-            data[5].Size = sizeof(double);
-            data[5].Reserved = 0;
-            data[6].DataPointer = (IntPtr)(&ThroughputRatio);
-            data[6].Size = sizeof(double);
-            data[6].Reserved = 0;
-            data[7].DataPointer = (IntPtr)(&Confidence);
-            data[7].Size = sizeof(double);
-            data[7].Reserved = 0;
-            data[8].DataPointer = (IntPtr)(&NewControlSetting);
-            data[8].Size = sizeof(double);
-            data[8].Reserved = 0;
-            data[9].DataPointer = (IntPtr)(&NewThreadWaveMagnitude);
-            data[9].Size = sizeof(ushort);
-            data[9].Reserved = 0;
-            data[10].DataPointer = (IntPtr)(&ClrInstanceID);
-            data[10].Size = sizeof(ushort);
-            data[10].Reserved = 0;
-            WriteEventCore(56, 11, data);
-        }
-
-        [Event(63, Level = EventLevel.Verbose, Message = Messages.IOEnqueue, Task = Tasks.ThreadPool, Opcode = Opcodes.IOEnqueue, Version = 0, Keywords = Keywords.ThreadingKeyword | Keywords.ThreadTransferKeyword)]
-        private unsafe void ThreadPoolIOEnqueue(
-            IntPtr NativeOverlapped,
-            IntPtr Overlapped,
-            bool MultiDequeues,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            int multiDequeuesInt = Convert.ToInt32(MultiDequeues); // bool maps to "win:Boolean", a 4-byte boolean
-
-            EventData* data = stackalloc EventData[4];
-            data[0].DataPointer = (IntPtr)(&NativeOverlapped);
-            data[0].Size = IntPtr.Size;
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&Overlapped);
-            data[1].Size = IntPtr.Size;
-            data[1].Reserved = 0;
-            data[2].DataPointer = (IntPtr)(&multiDequeuesInt);
-            data[2].Size = sizeof(int);
-            data[2].Reserved = 0;
-            data[3].DataPointer = (IntPtr)(&ClrInstanceID);
-            data[3].Size = sizeof(ushort);
-            data[3].Reserved = 0;
-            WriteEventCore(63, 4, data);
-        }
-
-        // TODO: This event is fired for minor compat with CoreCLR in this case. Consider removing this method and use
-        // FrameworkEventSource's thread transfer send/receive events instead at callers.
-        [NonEvent]
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        public void ThreadPoolIOEnqueue(RegisteredWaitHandle registeredWaitHandle)
-        {
-            if (IsEnabled(EventLevel.Verbose, Keywords.ThreadingKeyword | Keywords.ThreadTransferKeyword))
-            {
-                ThreadPoolIOEnqueue((IntPtr)registeredWaitHandle.GetHashCode(), IntPtr.Zero, registeredWaitHandle.Repeating);
-            }
-        }
-
-        [Event(64, Level = EventLevel.Verbose, Message = Messages.IO, Task = Tasks.ThreadPool, Opcode = Opcodes.IODequeue, Version = 0, Keywords = Keywords.ThreadingKeyword | Keywords.ThreadTransferKeyword)]
-        private unsafe void ThreadPoolIODequeue(
-            IntPtr NativeOverlapped,
-            IntPtr Overlapped,
-            ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            EventData* data = stackalloc EventData[3];
-            data[0].DataPointer = (IntPtr)(&NativeOverlapped);
-            data[0].Size = IntPtr.Size;
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&Overlapped);
-            data[1].Size = IntPtr.Size;
-            data[1].Reserved = 0;
-            data[2].DataPointer = (IntPtr)(&ClrInstanceID);
-            data[2].Size = sizeof(ushort);
-            data[2].Reserved = 0;
-            WriteEventCore(64, 3, data);
-        }
-
-        // TODO: This event is fired for minor compat with CoreCLR in this case. Consider removing this method and use
-        // FrameworkEventSource's thread transfer send/receive events instead at callers.
-        [NonEvent]
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        public void ThreadPoolIODequeue(RegisteredWaitHandle registeredWaitHandle)
-        {
-            if (IsEnabled(EventLevel.Verbose, Keywords.ThreadingKeyword | Keywords.ThreadTransferKeyword))
-            {
-                ThreadPoolIODequeue((IntPtr)registeredWaitHandle.GetHashCode(), IntPtr.Zero);
-            }
-        }
-
-        [Event(60, Level = EventLevel.Verbose, Message = Messages.WorkingThreadCount, Task = Tasks.ThreadPoolWorkingThreadCount, Opcode = EventOpcode.Start, Version = 0, Keywords = Keywords.ThreadingKeyword)]
-        public unsafe void ThreadPoolWorkingThreadCount(uint Count, ushort ClrInstanceID = DefaultClrInstanceId)
-        {
-            if (!IsEnabled(EventLevel.Verbose, Keywords.ThreadingKeyword))
-            {
-                return;
-            }
-
-            EventData* data = stackalloc EventData[2];
-            data[0].DataPointer = (IntPtr)(&Count);
-            data[0].Size = sizeof(uint);
-            data[0].Reserved = 0;
-            data[1].DataPointer = (IntPtr)(&ClrInstanceID);
-            data[1].Size = sizeof(ushort);
-            data[1].Reserved = 0;
-            WriteEventCore(60, 2, data);
-        }
-
-#pragma warning disable IDE1006 // Naming Styles
-        public static readonly PortableThreadPoolEventSource Log = new PortableThreadPoolEventSource();
-#pragma warning restore IDE1006 // Naming Styles
-    }
-}
